made chages for button and icons
