+++ src/App.tsx
@@ -41,6 +41,5 @@ function App() {
     </div>
   );
 }
-{/* <input onChange={(e) => debounce(()=>update(e.target.value), 500)} /> */ }
   // eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiJmNTg4Y2NmOC1jYjhkLTRjMzAtYjg2MC1lNmM3ZjQwZmNmYjIiLCJlbWFpbCI6ImFkYXJzaGJhbGlrYUBnbWFpbC5jb20ifQ.AMka3wzV3nXZA2OEKejEdL546VHpTIrP1GszqcbCbO8
 export default App;
