+++ src/components/UI/Buttons/Buttons.tsx
@@ -1,0 +1,24 @@
+import {BiJournalPlus, BiPatchPlus} from '../Icons/Icons';
+import "./Buttons.css";
+
+function FloatAddButton() {
+  return (
+    <div className='float-add-btn'><BiJournalPlus /></div>
+  )
+}
+
+
+
+function CreateButton(props) {
+    return (
+      <div className='create-btn'>Create Note <BiPatchPlus/></div>
+    )
+  }
+  
+  CreateButton.propTypes = {}
+  
+export {
+    CreateButton,
+    FloatAddButton
+};
+
