--- src/components/UI/Buttons/CreateButton/CreateButton.css
@@ -1,21 +1,0 @@
-.create-btn{
-    color: white;
-    width: fit-content;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    word-spacing: 0.1em;
-    margin: 0.5em;
-    padding: 0.5em;
-    border-radius: 0.5rem;
-    gap: 1rem;
-    font-size: 1.5em;
-    font-weight: 700;
-    background-color: red;
-    cursor: pointer;
-}
-
-
-.create-btn:hover{
-    transition: scale(1.05);
-}
