--- src/components/UI/Buttons/FloatAddButton/FloatAddButton.css
@@ -1,22 +1,0 @@
-.float-add-btn{
-    height: 4rem;
-    width: 4rem;
-    border-radius: 50%;
-    position: fixed;
-    bottom: 2rem;
-    right: 3rem;
-    background-color: var(--floatbtn-bg);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    color: burlywood;
-    font-size: 2rem;
-    aspect-ratio: 1;
-    cursor: pointer;
-}
-
-.float-add-btn:hover{
-    transform: scale(1.1);
-    background-color: rgba(--floatbtn-bg, 0.01);
-    color: white;
-}
