--- src/components/UI/Buttons/FloatAddButton/FloatAddButton.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import BiJournalPlus from '../../Icons/BiJournalPlus';
-import "./FloatAddButton.css";
-
-function FloatAddButton() {
-  return (
-    <div className='float-add-btn'><BiJournalPlus /></div>
-  )
-}
-
-export default FloatAddButton
