+++ src/components/common/AsideNavbar/AsideNavbar.tsx
@@ -1,9 +1,4 @@
-import React from 'react';
-import BiArchive from '../../UI/Icons/BiArchive';
-import BiBoxArrowRight from '../../UI/Icons/BiBoxArrowRight';
-import BiHouseHeart from '../../UI/Icons/BiHouseHeart';
-import BiPersonHearts from '../../UI/Icons/BiPersonHearts';
-import { BiTrash } from '../../UI/Icons/BiTrash';
+import { BiTrash, BiPersonHearts, BiHouseHeart, BiBoxArrowRight, BiArchive } from '../../UI/Icons/Icons';
 import "./AsideNavbar.css";
 
 function AsideNavbar() {
@@ -11,9 +6,9 @@ function AsideNavbar() {
         <div className='AsideNav'>
             <div>
                 <ul>
-                    <li><BiHouseHeart/>Home</li>
+                    <li><BiHouseHeart />Home</li>
                     <li><BiArchive />Archive</li>
-                    <li><BiTrash/> Trash</li>
+                    <li><BiTrash /> Trash</li>
                     <li><BiPersonHearts />Profile</li>
                 </ul>
             </div>
