+++ src/customhook/useDocumentationTitle.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function useDocumentationTitle() {
+  return (
+    <div>useDocumentationTitle</div>
+  )
+}
+
+export default useDocumentationTitle
