+++ src/index.tsx
@@ -1,11 +1,4 @@
-// import React from 'react';
-// // import ReactDOM from 'react-dom';
-// import ReactDOM from "react-dom/client";
-// import './index.css';
-// import App from './App';
-// import reportWebVitals from './reportWebVitals';
 import { BrowserRouter } from "react-router-dom";
-
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
@@ -22,12 +15,6 @@ root.render(
         </BrowserRouter>
   </React.StrictMode>
 );
-// ReactDOM.render(
-//   <React.StrictMode>
-//     <App />
-//   </React.StrictMode>,
-//   document.getElementById('root')
-// );
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
