+++ src/pages/HomePage/HomePage.tsx
@@ -1,25 +1,28 @@
+/* eslint-disable no-lone-blocks */
 import "./HomePage.css";
 import ReactQuill from 'react-quill';
-import CreateButton from "../../components/UI/Buttons/CreateButton/CreateButton";
 import { useReducer, useRef, useState } from "react";
 import { debounce } from "../../utils/Debounce";
 import DOMPurify from 'dompurify';
-import FloatAddButton from "../../components/UI/Buttons/FloatAddButton/FloatAddButton";
-import BiXCircle from "../../components/UI/Icons/BiXCircle";
+import { FloatAddButton, CreateButton } from "../../components/UI/Buttons/Buttons";
+import { BiXCircle } from "../../components/UI/Icons/Icons";
 
-const ContentDetail = (state, action) => { 
-  console.log(state);
-  console.log(action, action.data); 
-  if (action.type === "htmlbody") { 
-    return { ...state ,htmlbody :  action.data}
+const ContentDetail = (state, action) => {
+  switch (action.type) {
+    // eslint-disable-next-line no-lone-blocks
+    case "htmlbody": {
+      return { ...state, htmlbody: action.data }
+    };
+    case "title": {
+      return { ...state, title: action.data }
+    };
+    case "color": {
+      return { ...state, color: action.data }
+    };
+    default: {
+      return { ...state };
+    }
   }
-  if (action.type === "title") { 
-    return { ...state ,title : action.data}
-  }
-  if (action.type === "color") { 
-    return { ...state ,color : action.data}
-  }
-  return { ...state };
 }
 
 
@@ -31,11 +34,11 @@ export default function HomePage() {
       title: "",
       htmlbody: "",
       createdOn: "",
-      color :"",
+      color: "",
     });
 
-  
- 
+
+
   var modules = {
     toolbar: [
       [{ 'header': [1, 2, false] }],
@@ -43,9 +46,9 @@ export default function HomePage() {
       [{ 'list': 'ordered' }, { 'list': 'bullet' }, { 'indent': '-1' }, { 'indent': '+1' }],
       ['link', 'image'],
       ['clean']
-     ],
+    ],
   };
- 
+
   // var formats = [
   //   'header',
   //   'bold', 'italic', 'underline', 'strike', 'blockquote',
@@ -66,35 +69,35 @@ export default function HomePage() {
       <div>
         {showNote &&
           <div className="note-editor-container">
-            <div className="close-note" onClick={()=>setShowNote(false)}><BiXCircle /></div>
-          <input 
-            className="note-title-input"
-            placeholder="Title note ...."
-            onChange={(e) => debounce(() => noteDispatch({ type: "title", data: e.target.value }), 500)} />
-          <ReactQuill
-            theme="snow"
-            modules={modules}
-            // formats={formats}
-            placeholder={"Write Something....."}
-            // value={noteState.htmlbody}
-            onChange={(e) => debounce(() => noteDispatch({ type: "htmlbody", data : e }), 500)} />  
-          <div className="note-editor-action">
+            <div className="close-note" onClick={() => setShowNote(false)}><BiXCircle /></div>
+            <input
+              className="note-title-input"
+              placeholder="Title note ...."
+              onChange={(e) => debounce(() => noteDispatch({ type: "title", data: e.target.value }), 500)} />
+            <ReactQuill
+              theme="snow"
+              modules={modules}
+              // formats={formats}
+              placeholder={"Write Something....."}
+              // value={noteState.htmlbody}
+              onChange={(e) => debounce(() => noteDispatch({ type: "htmlbody", data: e }), 500)} />
+            <div className="note-editor-action">
               <p className="color-schema" dangerouslySetInnerHTML={{ __html: DOMPurify.sanitize(noteState.htmlbody) }}></p>
               <input
                 type="color" className="color-pallete"
-                onChange={(e) => debounce(() => noteDispatch({ type: "color", data: e.target.value }), 500)}/>
-            <span><CreateButton /></span>
-          </div>
+                onChange={(e) => debounce(() => noteDispatch({ type: "color", data: e.target.value }), 500)} />
+              <span><CreateButton /></span>
+            </div>
           </div>
         }
       </div>
-      
-     
+
+
       <div className="latest-notes-container">
         <h1>Latest Notes : </h1>
         <div></div>
       </div>
-      <span onClick={()=> setShowNote(true)}>
+      <span onClick={() => setShowNote(true)}>
         <FloatAddButton />
       </span>
     </div>
